docs: implement automatic API generation using `typedoc` (#2)
This PR fixes https://github.com/imballinst/count-up-down/issues/1.

## `typedoc` and `typedoc-plugin-markdown`

These are new dependencies for this PR. Using these 2 dependencies, we will be able to generate Markdown files from the TypeScript files. This is very useful because using one script command:

```
"generate:doc": "rm -rf api/docs && typedoc --plugin typedoc-plugin-markdown --readme none --out api/docs src/node src/browser src/common/*.ts"
```

We will be able to update the API documentation that is linked in GitHub.

## Folder structure changes

Flat folder structure was a bit awkward when it comes to documentation generation like this, especially because this library can be used for both Node and browser. As such, "categorization" using nested folder was necessary so that the separation is clear in the output folder structure.